renamed player package to party
